python/requirements.txt
agentmail>=0.4.0
anthropic>=0.91.0
python-dotenv>=1.0.0
certifi>=2024.0.0
aiohttp>=3.9.0
